udtw-server/pyproject.toml
[tool.poetry]
name = "udtw-server"
version = "0.1.0"
description = ""
authors = ["rebzzel <rebzzel@gmail.com>"]
packages = [
    {include = 'udtw', from = 'src'},
]

[tool.poetry.dependencies]
python = "^3.10"
Flask = {extras = ["async"], version = "^2.2.2"}
google-api-python-client = "^2.60.0"
google-auth-httplib2 = "^0.1.0"
google-auth-oauthlib = "^0.5.2"
psycopg2 = "^2.9.3"
requests = "^2.28.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
